+++ src/app/tools/gen-mapper/dialogs/people-group-dialog/country-picker/country-picker.component.spec.ts
@@ -1,0 +1,25 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { configureTestSuite } from 'ng-bullet';
+
+import { CountryPickerComponent } from './country-picker.component';
+
+describe.skip('CountryPickerComponent', () => {
+    let component: CountryPickerComponent;
+    let fixture: ComponentFixture<CountryPickerComponent>;
+
+    configureTestSuite(() => {
+        TestBed.configureTestingModule({
+            declarations: [CountryPickerComponent]
+        });
+    });
+
+    beforeEach(() => {
+        fixture = TestBed.createComponent(CountryPickerComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it('should create', () => {
+        expect(component).toBeTruthy();
+    });
+});